+++ src/commonMain/kotlin/org/eclipse/keyple/keyplelessdistributedlib/protocol/LogLevel.kt
@@ -9,12 +9,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ************************************************************************************** */
-package org.eclipse.keyple.distributed.protocol
+package org.eclipse.keyple.keyplelessdistributedlib.protocol
 
-sealed class KeypleResult<T> {
-  data class Success<T>(val data: T) : KeypleResult<T>()
-
-  data class Failure<T>(val error: KeypleError) : KeypleResult<T>()
+enum class LogLevel {
+  NONE,
+  INFO,
+  ERROR,
+  DEBUG
 }
-
-data class KeypleError(val statusCode: Int, val message: String)
